@@ -4,9 +4,10 @@ description: Demonstrates how to use the cronet_http plugin.
 publish_to: 'none'
 
 environment:
-  sdk: ">=2.17.5 <3.0.0"
+  sdk: ">=2.19.0 <3.0.0"
 
 dependencies:
+  cached_network_image: ^3.2.3
   cronet_http:
     path: ../
   cupertino_icons: ^1.0.2
@@ -15,13 +16,14 @@ dependencies:
   http: ^0.13.5
 
 dev_dependencies:
-  flutter_lints: ^1.0.0
+  dart_flutter_team_lints: ^1.0.0
   flutter_test:
     sdk: flutter
   http_client_conformance_tests:
     path: ../../http_client_conformance_tests/
   integration_test:
     sdk: flutter
+  test: ^1.23.1
 
 flutter:
   uses-material-design: true
