@@ -4,7 +4,7 @@ homepage: https://github.com/JonathanPeterCole/dart-cancellation-token-http/tree
 description: A fork of Dart's HTTP package with request cancellation using cancellation_token.
 
 environment:
-  sdk: '>=2.15.0 <3.0.0'
+  sdk: '>=2.19.0 <3.0.0'
 
 platforms:
   android:
@@ -19,13 +19,12 @@ dependencies:
   cancellation_token: ^1.5.0
   http_parser: ^4.0.0
   meta: ^1.3.0
-  path: ^1.8.0
 
 dev_dependencies:
+  dart_flutter_team_lints: ^1.0.0
   fake_async: ^1.2.0
   http_client_conformance_tests:
     path: ../http_client_conformance_tests/
-  lints: '>=1.0.0 <3.0.0'
   shelf: ^1.1.0
   stream_channel: ^2.1.0
   test: ^1.16.0
