@@ -6,13 +6,14 @@ publish_to: none
 repository: https://github.com/dart-lang/http/tree/master/pkgs/http_client_conformance_tests
 
 environment:
-  sdk: '>=2.14.0 <3.0.0'
+  sdk: '>=2.19.0 <3.0.0'
 
 dependencies:
   async: ^2.8.2
   cancellation_token_http: ^1.1.0
+  dart_style: ^2.2.3
+  stream_channel: ^2.1.1
   test: ^1.21.2
 
 dev_dependencies:
-  dart_style: ^2.2.3
-  lints: ^1.0.0
+  dart_flutter_team_lints: ^1.0.0
