@@ -2,7 +2,6 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-import 'package:cupertino_http/cupertino_client.dart';
 import 'package:cupertino_http/cupertino_http.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:http_client_conformance_tests/http_client_conformance_tests.dart';
@@ -12,12 +11,12 @@ void main() {
   IntegrationTestWidgetsFlutterBinding.ensureInitialized();
 
   group('defaultSessionConfiguration', () {
-    testAll(CupertinoClient.defaultSessionConfiguration(),
+    testAll(CupertinoClient.defaultSessionConfiguration,
         canStreamRequestBody: false);
   });
   group('fromSessionConfiguration', () {
     final config = URLSessionConfiguration.ephemeralSessionConfiguration();
-    testAll(CupertinoClient.fromSessionConfiguration(config),
-        canStreamRequestBody: false);
+    testAll(() => CupertinoClient.fromSessionConfiguration(config),
+        canStreamRequestBody: false, canWorkInIsolates: false);
   });
 }
