@@ -5,6 +5,7 @@
 import 'dart:io';
 
 import 'package:cupertino_http/cupertino_http.dart';
+import 'package:flutter/foundation.dart';
 import 'package:integration_test/integration_test.dart';
 import 'package:test/test.dart';
 
@@ -70,7 +71,13 @@ void testURLSessionTask(
         });
       final session = URLSession.sharedSession();
       task = session.dataTaskWithRequest(
-          URLRequest.fromUrl(Uri.parse('http://localhost:${server.port}')))
+          MutableURLRequest.fromUrl(
+              Uri.parse('http://localhost:${server.port}/mypath'))
+            ..httpMethod = 'POST'
+            ..httpBody = Data.fromUint8List(Uint8List.fromList([1, 2, 3])))
+        ..prefersIncrementalDelivery = false
+        ..priority = 0.2
+        ..taskDescription = 'my task description'
         ..resume();
 
       while (task.state != URLSessionTaskState.urlSessionTaskStateCompleted) {
@@ -83,8 +90,24 @@ void testURLSessionTask(
       server.close();
     });
 
+    test('priority', () async {
+      expect(task.priority, inInclusiveRange(0.19, 0.21));
+    });
+
+    test('current request', () async {
+      expect(task.currentRequest!.url!.path, '/mypath');
+    });
+
+    test('original request', () async {
+      expect(task.originalRequest!.url!.path, '/mypath');
+    });
+
     test('has response', () async {
       expect(task.response, isA<HTTPURLResponse>());
+    });
+
+    test('no error', () async {
+      expect(task.error, null);
     });
 
     test('countOfBytesExpectedToReceive - no content length set', () async {
@@ -95,8 +118,99 @@ void testURLSessionTask(
       expect(task.countOfBytesReceived, 11);
     });
 
+    test('countOfBytesExpectedToSend', () async {
+      expect(task.countOfBytesExpectedToSend, 3);
+    });
+
+    test('countOfBytesSent', () async {
+      expect(task.countOfBytesSent, 3);
+    });
+
+    test('taskDescription', () {
+      expect(task.taskDescription, 'my task description');
+    });
+
     test('taskIdentifier', () {
       task.taskIdentifier; // Just verify that there is no crash.
+    });
+
+    test('prefersIncrementalDelivery', () {
+      expect(task.prefersIncrementalDelivery, false);
+    });
+
+    test('toString', () {
+      task.toString(); // Just verify that there is no crash.
+    });
+  });
+
+  group('task failed', () {
+    late URLSessionTask task;
+    setUp(() async {
+      final session = URLSession.sharedSession();
+      task = session.dataTaskWithRequest(
+          MutableURLRequest.fromUrl(Uri.parse('http://notarealserver')))
+        ..resume();
+
+      while (task.state != URLSessionTaskState.urlSessionTaskStateCompleted) {
+        // Let the event loop run.
+        await Future<void>(() {});
+      }
+    });
+    tearDown(() {
+      task.cancel();
+    });
+
+    test('no response', () async {
+      expect(task.response, null);
+    });
+
+    test('no error', () async {
+      expect(task.error!.code, -1003); // CannotFindHost
+    });
+
+    test('toString', () {
+      task.toString(); // Just verify that there is no crash.
+    });
+  });
+
+  group('task redirect', () {
+    late HttpServer server;
+    late URLSessionTask task;
+    setUp(() async {
+      // The task will request http://localhost:XXX/launch, which will be
+      // redirected to http://localhost:XXX/landed.
+      server = (await HttpServer.bind('localhost', 0))
+        ..listen((request) async {
+          await request.drain<void>();
+          if (request.requestedUri.path != '/landed') {
+            await request.response.redirect(Uri(path: '/landed'));
+            return;
+          }
+          request.response.headers.set('Content-Type', 'text/plain');
+          request.response.write('Hello World');
+          await request.response.close();
+        });
+      final session = URLSession.sharedSession();
+      task = session.dataTaskWithRequest(MutableURLRequest.fromUrl(
+          Uri.parse('http://localhost:${server.port}/launch')))
+        ..resume();
+
+      while (task.state != URLSessionTaskState.urlSessionTaskStateCompleted) {
+        // Let the event loop run.
+        await Future<void>(() {});
+      }
+    });
+    tearDown(() {
+      task.cancel();
+      server.close();
+    });
+
+    test('current request', () async {
+      expect(task.currentRequest!.url!.path, '/landed');
+    });
+
+    test('original request', () async {
+      expect(task.originalRequest!.url!.path, '/launch');
     });
 
     test('toString', () {
