@@ -5,7 +5,8 @@
 import 'dart:convert';
 import 'dart:io';
 
-import 'package:cupertino_http/cupertino_client.dart';
+import 'package:cached_network_image/cached_network_image.dart';
+import 'package:cupertino_http/cupertino_http.dart';
 import 'package:flutter/material.dart';
 import 'package:http/http.dart';
 
@@ -121,7 +122,11 @@ class _BookListState extends State<BookList> {
         itemBuilder: (context, index) => Card(
           key: ValueKey(widget.books[index].title),
           child: ListTile(
-            leading: Image.network(widget.books[index].imageUrl),
+            leading: CachedNetworkImage(
+                placeholder: (context, url) =>
+                    const CircularProgressIndicator(),
+                imageUrl:
+                    widget.books[index].imageUrl.replaceFirst('http', 'https')),
             title: Text(widget.books[index].title),
             subtitle: Text(widget.books[index].description),
           ),
