@@ -11,6 +11,10 @@ import 'package:cancellation_token/cancellation_token.dart';
 import 'http.dart';
 
 /// An HTTP client wrapper that automatically retries failing requests.
+///
+/// NOTE: [RetryClient] makes a copy of the request data in order to support
+/// resending it. This can cause a lot of memory usage when sending a large
+/// [StreamedRequest].
 class RetryClient extends BaseClient {
   /// The wrapped client.
   final Client _inner;
@@ -65,8 +69,8 @@ class RetryClient extends BaseClient {
     RangeError.checkNotNegative(_retries, 'retries');
   }
 
-  /// Like [RetryClient], but with a pre-computed list of [delays] between each
-  /// retry.
+  /// Like [RetryClient.new], but with a pre-computed list of [delays]
+  /// between each retry.
   ///
   /// This will retry a request at most `delays.length` times, using each delay
   /// in order. It will wait for `delays[0]` after the initial request,
