@@ -5,14 +5,14 @@
 import 'dart:io';
 
 import 'package:http_parser/http_parser.dart';
-import 'package:path/path.dart' as p;
 
 import 'byte_stream.dart';
 import 'multipart_file.dart';
 
 Future<MultipartFile> multipartFileFromPath(String field, String filePath,
     {String? filename, MediaType? contentType}) async {
-  filename ??= p.basename(filePath);
+  late var segments = Uri.file(filePath).pathSegments;
+  filename ??= segments.isEmpty ? '' : segments.last;
   var file = File(filePath);
   var length = await file.length();
   var stream = ByteStream(file.openRead());
