@@ -2,7 +2,6 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-import 'base_request.dart';
 import 'base_response.dart';
 import 'byte_stream.dart';
 import 'utils.dart';
@@ -18,19 +17,12 @@ class StreamedResponse extends BaseResponse {
   /// Creates a new streaming response.
   ///
   /// [stream] should be a single-subscription stream.
-  StreamedResponse(Stream<List<int>> stream, int statusCode,
-      {int? contentLength,
-      BaseRequest? request,
-      Map<String, String> headers = const {},
-      bool isRedirect = false,
-      bool persistentConnection = true,
-      String? reasonPhrase})
-      : stream = toByteStream(stream),
-        super(statusCode,
-            contentLength: contentLength,
-            request: request,
-            headers: headers,
-            isRedirect: isRedirect,
-            persistentConnection: persistentConnection,
-            reasonPhrase: reasonPhrase);
+  StreamedResponse(Stream<List<int>> stream, super.statusCode,
+      {super.contentLength,
+      super.request,
+      super.headers,
+      super.isRedirect,
+      super.persistentConnection,
+      super.reasonPhrase})
+      : stream = toByteStream(stream);
 }
