@@ -3,16 +3,16 @@
 // BSD-style license that can be found in the LICENSE file.
 
 @TestOn('browser')
+library;
 
 import 'package:cancellation_token_http/browser_client.dart';
 import 'package:http_client_conformance_tests/http_client_conformance_tests.dart';
 import 'package:test/test.dart';
 
 void main() {
-  final client = BrowserClient();
-
-  testAll(client,
+  testAll(BrowserClient.new,
       redirectAlwaysAllowed: true,
       canStreamRequestBody: false,
-      canStreamResponseBody: false);
+      canStreamResponseBody: false,
+      canWorkInIsolates: false);
 }
