@@ -3,13 +3,12 @@
 // BSD-style license that can be found in the LICENSE file.
 
 @TestOn('vm')
+library;
 
 import 'package:cancellation_token_http/io_client.dart';
 import 'package:http_client_conformance_tests/http_client_conformance_tests.dart';
 import 'package:test/test.dart';
 
 void main() {
-  final client = IOClient();
-
-  testAll(client);
+  testAll(IOClient.new);
 }
