@@ -15,6 +15,6 @@ class MyHttpClient extends BaseClient {
 
 void main() {
   group('client conformance tests', () {
-    testAll(MyHttpClient());
+    testAll(MyHttpClient.new);
   });
 }
